Reimplement persistence id as BigInteger
This allows us to use MOngoDB object ids (which are 12 bytes)

+++ src/main/java/uk/ac/diamond/daq/persistence/logging/ConfigurationLogService.java
@@ -3,10 +3,11 @@ package uk.ac.diamond.daq.persistence.logging;
 import uk.ac.diamond.daq.persistence.data.PersistableItem;
 import uk.ac.diamond.daq.persistence.service.PersistenceException;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface ConfigurationLogService {
-    List<LogToken> listChanges(long persistenceId, Class<?> clazz) throws PersistenceException;
+    List<LogToken> listChanges(BigInteger persistenceId, Class<?> clazz) throws PersistenceException;
 
     List<LogToken> listChanges(PersistableItem item) throws PersistenceException;
