+++ src/main/java/uk/ac/diamond/daq/persistence/service/PersistenceService.java
@@ -2,6 +2,7 @@ package uk.ac.diamond.daq.persistence.service;
 
 import uk.ac.diamond.daq.persistence.data.PersistableItem;
 
+import java.math.BigInteger;
 import java.util.List;
 import java.util.Map;
 
@@ -14,7 +15,7 @@ public interface PersistenceService {
 
     void save(PersistableItem item) throws PersistenceException;
 
-    void delete(long persistenceId);
+    void delete(BigInteger persistenceId);
 
     void delete(PersistableItem item);
 
@@ -24,9 +25,9 @@ public interface PersistenceService {
     <T extends PersistableItem> SearchResult get(Map<String, String> searchParameters, Class<T> clazz)
                     throws PersistenceException;
 
-    <T extends PersistableItem> T get(long persistenceId, Class<T> clazz) throws PersistenceException;
+    <T extends PersistableItem> T get(BigInteger persistenceId, Class<T> clazz) throws PersistenceException;
 
-    List<Long> getVersions(long persistenceId);
+    List<Long> getVersions(BigInteger persistenceId);
 
-    <T extends PersistableItem> T getArchive(long persistenceId, long version, Class<T> clazz) throws PersistenceException;
+    <T extends PersistableItem> T getArchive(BigInteger persistenceId, long version, Class<T> clazz) throws PersistenceException;
 }
