+++ src/main/java/uk/ac/diamond/daq/persistence/service/SearchResultRow.java
@@ -1,20 +1,21 @@
 package uk.ac.diamond.daq.persistence.service;
 
+import java.math.BigInteger;
 import java.util.Map;
 
 public class SearchResultRow {
-    private long persistenceId;
+    private BigInteger persistenceId;
     private long version;
 
     private Map<SearchResultHeading, String> values;
 
-    public SearchResultRow(long persistenceId, long version, Map<SearchResultHeading, String> values) {
+    public SearchResultRow(BigInteger persistenceId, long version, Map<SearchResultHeading, String> values) {
         this.persistenceId = persistenceId;
         this.version = version;
         this.values = values;
     }
 
-    public long getPersistenceId() {
+    public BigInteger getPersistenceId() {
         return persistenceId;
     }
